Extract just problematic test from BinaryWriter.Encoding tests
